requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
scipy>=1.11
python-dotenv>=1.0
psycopg2-binary>=2.9
boto3>=1.34
sagemaker>=2.200
streamlit>=1.30
pyarrow>=14.0
pytest>=7.4
